# Speedup CI tests