--- a/x/ibc-hooks/ibc_middleware_test.go
+++ b/x/ibc-hooks/ibc_middleware_test.go
@@ -3,8 +3,9 @@ package ibc_hooks_test
 import (
 	"encoding/json"
 	"fmt"
-	ibc_hooks "github.com/osmosis-labs/osmosis/v13/x/ibc-hooks"
 	"testing"
+
+	ibc_hooks "github.com/osmosis-labs/osmosis/v13/x/ibc-hooks"
 
 	"github.com/osmosis-labs/osmosis/v13/osmoutils"
 
@@ -45,6 +46,10 @@ func (suite *HooksTestSuite) SetupTest() {
 	suite.chainB = &osmosisibctesting.TestChain{
 		TestChain: suite.coordinator.GetChain(ibctesting.GetChainID(2)),
 	}
+	err := suite.chainA.MoveEpochsToTheFuture()
+	suite.Require().NoError(err)
+	err = suite.chainB.MoveEpochsToTheFuture()
+	suite.Require().NoError(err)
 	suite.path = NewTransferPath(suite.chainA, suite.chainB)
 	suite.coordinator.Setup(suite.path)
 }