--- a/x/ibc-rate-limit/testutil/chain.go
+++ b/x/ibc-rate-limit/testutil/chain.go
@@ -10,9 +10,10 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ibctesting "github.com/cosmos/ibc-go/v3/testing"
 	"github.com/cosmos/ibc-go/v3/testing/simapp/helpers"
-	"github.com/osmosis-labs/osmosis/v13/app"
 	abci "github.com/tendermint/tendermint/abci/types"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	"github.com/osmosis-labs/osmosis/v13/app"
 )
 
 type TestChain struct {
@@ -85,14 +86,18 @@ func SignAndDeliver(
 }
 
 // Move epochs to the future to avoid issues with minting
-func (chain *TestChain) MoveEpochsToTheFuture() {
+func (chain *TestChain) MoveEpochsToTheFuture() error {
 	epochsKeeper := chain.GetOsmosisApp().EpochsKeeper
 	ctx := chain.GetContext()
 	for _, epoch := range epochsKeeper.AllEpochInfos(ctx) {
 		epoch.StartTime = ctx.BlockTime().Add(time.Hour * 24 * 30)
 		epochsKeeper.DeleteEpochInfo(chain.GetContext(), epoch.Identifier)
-		_ = epochsKeeper.AddEpochInfo(ctx, epoch)
+		err := epochsKeeper.AddEpochInfo(ctx, epoch)
+		if err != nil {
+			return err
+		}
 	}
+	return nil
 }
 
 // GetOsmosisApp returns the current chain's app as an OsmosisApp