<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## What is the purpose of the change

Does basic changes to speedup tests, driven by `x/ibc-rate-limit`, but should have more broad impact.

Locally `13.943s -> 2.670s`, CI: `86.922s -> 10.370s`. As a side effect:
IBC-hooks `107.357s -> 3.827s`

- Disable epochs on both chains in ibc-rate-limit tests (big win)
- Make epochs not allocate high amounts of RAM for events unless theres at least 10 gauges (small win)
- Remove invariants from our testing setups (small win, across many tests)

## Testing and Verifying

All tests still pass.

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? no
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? (not applicable   /   specification N/A